src/requirements.txt
psycopg2==2.8.2
bs4==0.0.1
flask==1.0.2
Jinja2==2.10
flask-wtf==0.14.2
uWSGI==2.0.18
MarkupSafe==1.1.1
Flask-SQLAlchemy==2.4.0
Flask-Bcrypt==0.7.1